Allow extending datasets not yet staged (#123)
* allow extending datasets not yet staged

* wip

* define and use DatasetCardButton component

* move dataset card related components into DatasetCard folder
