+++ triton-client/package.json
@@ -31,6 +31,7 @@
     },
     "dependencies": {
         "@ant-design/charts": "^2.1.2",
+        "@ant-design/icons": "^5.5.1",
         "@antv/g2": "^5.2.2",
         "antd": "^5.18.0",
         "lodash": "^4.17.21",
